694 lab 665 lab 653 add view to update job flow (#696)

+++ internal/bacalhau/bacalhau_test.go
@@ -19,8 +19,8 @@ func TestCreateBacalhauJob(t *testing.T) {
 	if err != nil {
 		t.Fatalf(fmt.Sprint(err))
 	}
-	if job.Spec.Timeout != float64(timeOut) {
-		t.Errorf("got = %f; wanted %d", job.Spec.Timeout, timeOut)
+	if job.Spec.Timeout != int64(timeOut) {
+		t.Errorf("got = %d; wanted %d", job.Spec.Timeout, timeOut)
 	}
 	if job.Spec.Resources.Memory != memory {
 		t.Errorf("got = %s; wanted %s", job.Spec.Resources.Memory, memory)
